HRIS-326 [FE] Modify the remarks section on overtime/leave/approve/timeIn/timeOut/latetimeIn to cater emojis
