+++ client/src/components/molecules/AddNewLeaveModal/LeaveTab.tsx
@@ -14,6 +14,7 @@ import { X, Plus, Save, User, Minus, Coffee, FileText, Calendar, RefreshCcw } fr
 import TextField from './../TextField'
 import useLeave from '~/hooks/useLeave'
 import useProject from '~/hooks/useProject'
+import { Emoji } from '~/utils/types/emoji'
 import Input from '~/components/atoms/Input'
 import useUserQuery from '~/hooks/useUserQuery'
 import { Roles } from '~/utils/constants/roles'
@@ -21,6 +22,7 @@ import { queryClient } from '~/lib/queryClient'
 import SpinnerIcon from '~/utils/icons/SpinnerIcon'
 import { NewLeaveSchema } from '~/utils/validation'
 import { LeaveType } from '~/utils/types/leaveTypes'
+import EmojiPopoverPicker from './../EmojiPopoverPicker'
 import { User as UserType } from '~/utils/types/userTypes'
 import Button from '~/components/atoms/Buttons/ButtonAction'
 import { NewLeaveFormValues } from '~/utils/types/formValues'
@@ -94,10 +96,11 @@ const LeaveTab: FC<Props> = ({ isOpen, closeModal }): JSX.Element => {
 
   const {
     reset,
+    watch,
     control,
     register,
+    setValue,
     handleSubmit,
-    watch,
     formState: { errors, isSubmitting }
   } = useForm<NewLeaveFormValues>({
     mode: 'onTouched',
@@ -223,6 +226,9 @@ const LeaveTab: FC<Props> = ({ isOpen, closeModal }): JSX.Element => {
 
   // Remove Date
   const handleRemoveProject = (index: number): void => projectRemove(index)
+
+  const handleEmojiSelect = (emoji: Emoji): void =>
+    setValue('reason', watch('reason') + emoji.native)
 
   return (
     <Tab.Panel className="focus:outline-none">
@@ -559,19 +565,29 @@ const LeaveTab: FC<Props> = ({ isOpen, closeModal }): JSX.Element => {
 
           {/* Reason for leave Field */}
           <section className="col-span-2">
-            <TextField title="Reason for leave" Icon={FileText} isRequired>
-              <ReactTextareaAutosize
-                id="reason"
-                {...register('reason')}
-                className={classNames(
-                  'text-area-auto-resize pl-12',
-                  errors?.reason !== null && errors.reason !== undefined
-                    ? 'border-rose-500 ring-rose-500'
-                    : ''
-                )}
-                disabled={isSubmitting}
-              />
-            </TextField>
+            <div className="relative">
+              <TextField title="Reason for leave" Icon={FileText} isRequired>
+                <ReactTextareaAutosize
+                  id="reason"
+                  {...register('reason')}
+                  className={classNames(
+                    'text-area-auto-resize min-h-[13vh] pl-12 pb-8',
+                    errors?.reason !== null && errors.reason !== undefined
+                      ? 'border-rose-500 ring-rose-500'
+                      : ''
+                  )}
+                  disabled={isSubmitting}
+                />
+                <div className="absolute bottom-1 left-11">
+                  <EmojiPopoverPicker
+                    {...{
+                      handleEmojiSelect,
+                      panelPosition: 'left-0 bottom-8'
+                    }}
+                  />
+                </div>
+              </TextField>
+            </div>
             {errors.reason !== null && errors.reason !== undefined && (
               <span className="error text-[10px]">{errors.reason?.message}</span>
             )}
