+++ client/src/components/organisms/ViewDetailsDrawer/index.tsx
@@ -4,8 +4,8 @@ import classNames from 'classnames'
 import { useRouter } from 'next/router'
 import { Calendar, Clock, X } from 'react-feather'
 
-import Text from '~/components/atoms/Text'
 import UploadedFiles from './UploadFiles'
+import Text from '~/components/atoms/Text'
 import Avatar from '~/components/atoms/Avatar'
 import { IMedia } from '~/utils/types/timeEntryTypes'
 import handleImageError from '~/utils/handleImageError'
